benches: add a speed benchmark // rename the existing to overhead
