+++ Cargo.toml
@@ -18,6 +18,7 @@ exclude = [ "misc/lc3-tm4c", "misc/serial" ]
 [package] # TODO
 name = "lc3-utp"
 version = "0.0.0"
+autobenches = false
 
 [dev-dependencies]
 lc3-baseline-sim = { path = "baseline-sim", version = "0.1.0", default-features = false }
@@ -35,6 +36,9 @@ lazy_static = "1.4.0"
 name = "speed"
 harness = false
 
+[[bench]]
+name = "overhead"
+harness = false
 
 [profile.bench]
 opt-level = 3
@@ -43,6 +47,8 @@ rpath = false
 lto = "fat"
 debug-assertions = false
 codegen-units = 1
+# target-cpu = "native"
+# panic = "abort"
 incremental = false
 overflow-checks = false
